tests/python/requirements.txt
pandas==1.0.5
pytest==6.0.1
requests<2.22,>=2
dataiku-api-client
git+git://github.com/dataiku/dataiku-plugin-tests-utils.git@master#egg=dataiku-plugin-tests-utils